fix: resolve PR review P1 blockers and P2 suggestions
- BatchDecode: partial scheduling + 100ms flush timeout + set reserve

- ROCm/CUDA sampling: clone probs, per-request seed, gather+log cum_log_probs

- VIT: round-robin addrs, proxy timeout fallback, URL singleflight

- MoriEP: shmem finalize, local_world_size, finalize debug log

- Quant: MXFP4 stacked MoE keys, fp4_moe_op pickle, fake_balance dtype check

- MM: input length/batch output validation, extra_input count check

- Renderers: preprocess_config for llama3/kimi_k25, empty TensorPB shape/dtype

- Qwen2-VL/Qwen3VL: max_seqlen once, CPU locs reuse

- Kernels: FlyDSL sentinel cache, remap fp32 cast, PureDP RS buffer cache

- BeamSearch: default large-beam boundary test

+++ rtp_llm/models_py/triton_kernels/moe/remap_local_ids_kernel.py
@@ -19,7 +19,7 @@ def _remap_to_local_ids_kernel(
     mask = offsets < N
 
     ids = tl.load(ids_ptr + offsets, mask=mask, other=0)
-    weights = tl.load(weights_ptr + offsets, mask=mask, other=0.0)
+    weights = tl.load(weights_ptr + offsets, mask=mask, other=0.0).to(tl.float32)
 
     is_local = (ids >= local_start) & (ids < local_end)
     local_ids = tl.where(is_local, ids - local_start, 0)
@@ -76,7 +76,7 @@ def remap_to_local_ids(
 
     _remap_to_local_ids_kernel[grid](
         dispatch_ids.view(-1),
-        dispatch_weights.to(torch.float32).view(-1),
+        dispatch_weights.view(-1),
         local_ids.view(-1),
         local_weights.view(-1),
         local_start,
